将 Ticket 改名成了 Access Token。准备接下来做OAuth的认证 Fix:DAO的JDBC实现中，错误的使用queryForObject。没有考虑它在处理没有找到数据时会抛出异常。 准备用户认证系统。

+++ steps-server/data-access-jdbc/src/main/resources/db-script/user.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS User(
-	`id` bigint(64) NOT NULL,
+	`id` serial NOT NULL,
 	`name` varchar(128) NOT NULL,
 	PRIMARY KEY(`id`),
 	UNIQUE KEY `idx_user_name` (`name`) USING HASH
